requirements.txt
confluent-kafka
requests
python-dotenv
elasticsearch>=8.0.0,<9.0.0
openai
gradio>=4.0.0
